update
change menu, inform
