+++ client-customer/src/App.css
@@ -63,7 +63,7 @@ form.search {
   margin: 0px;
 }
 input.keyword {
-  width: 200px;
+  width: 500px;
 }
 /* datatable.css */
 .datatable {
