+++ client-customer/src/components/InformComponent.js
@@ -1,6 +1,8 @@
 import React, { Component } from 'react';
 import { Link } from 'react-router-dom';
 import MyContext from '../contexts/MyContext';
+import withRouter from '../utils/withRouter';
+import axios from 'axios';
 
 import Container from 'react-bootstrap/Container';
 import Nav from 'react-bootstrap/Nav';
@@ -8,24 +10,30 @@ import Navbar from 'react-bootstrap/Navbar';
 import NavDropdown from 'react-bootstrap/NavDropdown';
 class Inform extends Component {
   static contextType = MyContext;
+  constructor(props) {
+    super(props);
+    this.state = {
+      categories: [],
+      txtKeyword: ''
+    };
+  }
   render() {
+    const cates = this.state.categories.map((item) => {
+      return (
+        // <li key={item._id} className="menu"><Link to={'/product/category/' + item._id}>{item.name}</Link></li>
+        <NavDropdown.Item key={item._id}><Link to={'/product/category/' + item._id} className='non'>{item.name}</Link></NavDropdown.Item>
+      );
+    });
     return (
       <Navbar expand="lg" className="bg-body-tertiary">
         <Container>
           <Navbar.Toggle aria-controls="basic-navbar-nav" />
           <Navbar.Collapse id="basic-navbar-nav">
-            <Nav className="me-auto">
-              {this.context.token === '' ?
-                <Nav.Item>
-                  <Link to='/login' className='non'>Login</Link> | <Link to='/signup' className='non'>Sign-up</Link> | <Link to='/active' className='non'>Active</Link>
-                </Nav.Item>
-                :
-                <NavDropdown title={this.context.customer.name} id="basic-nav-dropdown">
-                  <NavDropdown.Item><Link to='/myprofile' className='non'>My profile</Link></NavDropdown.Item>
-                  <NavDropdown.Item><Link to='/myorders' className='non'>My orders</Link></NavDropdown.Item>
-                  <NavDropdown.Item><Link to='/home' className='non' onClick={() => this.lnkLogoutClick()}>Logout</Link></NavDropdown.Item>
-                </NavDropdown>
-              }
+          <Nav className="me-auto">
+              <Nav.Link><Link to='/' className='non'>Home</Link></Nav.Link>
+              <NavDropdown title="Category" id="basic-nav-dropdown">
+                {cates}
+              </NavDropdown>
             </Nav>
             <Nav className="float-right">
               <Nav.Item><Link to='/mycart' className='non'>My cart</Link> have <b>{this.context.mycart.length}</b> items</Nav.Item>
@@ -35,11 +43,15 @@ class Inform extends Component {
       </Navbar>
     );
   }
-  // event-handlers
-  lnkLogoutClick() {
-    this.context.setToken('');
-    this.context.setCustomer(null);
-    this.context.setMycart([]);
+  componentDidMount() {
+    this.apiGetCategories();
+  }
+  // apis
+  apiGetCategories() {
+    axios.get('/api/customer/categories').then((res) => {
+      const result = res.data;
+      this.setState({ categories: result });
+    });
   }
 }
-export default Inform;
+export default withRouter(Inform);
