+++ client-customer/src/components/MenuComponent.js
@@ -1,7 +1,7 @@
-import axios from 'axios';
 import React, { Component } from 'react';
 import { Link } from 'react-router-dom';
 import withRouter from '../utils/withRouter';
+import MyContext from '../contexts/MyContext';
 
 import Container from 'react-bootstrap/Container';
 import Nav from 'react-bootstrap/Nav';
@@ -11,20 +11,14 @@ import Form from 'react-bootstrap/Form';
 import Button from 'react-bootstrap/Button';
 
 class Menu extends Component {
+  static contextType = MyContext;
   constructor(props) {
     super(props);
     this.state = {
-      categories: [],
-      txtKeyword: ''
+      txtKeyword: '',
     };
   }
   render() {
-    const cates = this.state.categories.map((item) => {
-      return (
-        // <li key={item._id} className="menu"><Link to={'/product/category/' + item._id}>{item.name}</Link></li>
-        <NavDropdown.Item key={item._id}><Link to={'/product/category/' + item._id} className='non'>{item.name}</Link></NavDropdown.Item>
-      );
-    });
     return (
       <Navbar expand="lg" className="bg-body-tertiary">
         <Container>
@@ -40,35 +34,34 @@ class Menu extends Component {
           </Navbar.Brand>
           <Navbar.Toggle aria-controls="basic-navbar-nav" />
           <Navbar.Collapse id="basic-navbar-nav">
-            <Nav className="me-auto">
-              <Nav.Link><Link to='/' className='non'>Home</Link></Nav.Link>
-              <NavDropdown title="Category" id="basic-nav-dropdown">
-                {cates}
-              </NavDropdown>
+            <Nav className="me-auto"></Nav> {/* khoảng trống*/}
+            <Form className="d-flex align-middle">
+              <Form.Control
+                type="search"
+                placeholder="Search"
+                className="keyword me-2 flex-grow-2" value={this.state.txtKeyword} onChange={(e) => { this.setState({ txtKeyword: e.target.value }) }}
+                aria-label="Search"
+                />
+              <Button type="submit" value="SEARCH" variant="outline-success" onClick={(e) => this.btnSearchClick(e)}>Search</Button>
+            </Form>
+            <Nav className="me-auto"></Nav> {/* khoảng trống*/}
+            <Nav className="float-right">
+              {this.context.token === '' ?
+                <Nav.Item>
+                  <Link to='/login' className='non'>Login</Link> | <Link to='/signup' className='non'>Sign-up</Link> | <Link to='/active' className='non'>Active</Link>
+                </Nav.Item>
+                :
+                <NavDropdown title={this.context.customer.name} id="basic-nav-dropdown">
+                  <NavDropdown.Item><Link to='/myprofile' className='non'>My profile</Link></NavDropdown.Item>
+                  <NavDropdown.Item><Link to='/myorders' className='non'>My orders</Link></NavDropdown.Item>
+                  <NavDropdown.Item><Link to='/home' className='non' onClick={() => this.lnkLogoutClick()}>Logout</Link></NavDropdown.Item>
+                </NavDropdown>
+              }
             </Nav>
           </Navbar.Collapse>
-          <Form className="d-flex">
-            <Form.Control
-              type="search"
-              placeholder="Search"
-              className="keyword me-2 flex-grow-2" value={this.state.txtKeyword} onChange={(e) => { this.setState({ txtKeyword: e.target.value }) }}
-              aria-label="Search"
-            />
-            <Button type="submit" value="SEARCH" variant="outline-success" onClick={(e) => this.btnSearchClick(e)}>Search</Button>
-          </Form>
         </Container>
       </Navbar>
     );
-  }
-  componentDidMount() {
-    this.apiGetCategories();
-  }
-  // apis
-  apiGetCategories() {
-    axios.get('/api/customer/categories').then((res) => {
-      const result = res.data;
-      this.setState({ categories: result });
-    });
   }
   // event-handlers
   btnSearchClick(e) {
@@ -77,8 +70,15 @@ class Menu extends Component {
     if (keyword !== '') {
       this.props.navigate('/product/search/' + this.state.txtKeyword);
     } else {
-      
+      //not found return to home page
+      this.props.navigate('/home');
     }
+  }
+  
+  lnkLogoutClick() {
+    this.context.setToken('');
+    this.context.setCustomer(null);
+    this.context.setMycart([]);
   }
 }
 export default withRouter(Menu);
